requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.7.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.5
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.7
Django==4.1.10
django-stubs==4.2.3
django-stubs-ext==4.2.2
django-webtest==1.9.10
djangorestframework==3.14.0
djangorestframework-stubs==3.14.2
drf-yasg==1.21.6
filelock==3.12.2
flake8==6.0.0
identify==2.5.24
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
model-bakery==1.12.0
mypy==1.4.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pre-commit==3.3.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.24.0
waitress==2.1.2
WebOb==1.8.7
WebTest==3.0.0
